+++ Front-End/src/components/friends/FriendsList.tsx
@@ -1,0 +1,81 @@
+import React, { FormEvent, useContext, useState } from 'react';
+import AuthContext from '../../store/AuthContext';
+import { Link } from "react-router-dom";
+import '../../style/Friends.css';
+import MyAvatar from '../user/Avatar';
+import { ListItem } from '@mui/material';
+import { ListItemAvatar } from '@mui/material';
+import { List } from '@mui/material';
+import ChatBubbleIcon from '@mui/icons-material/ChatBubble';
+import PersonAddIcon from '@mui/icons-material/PersonAdd';
+import { Snackbar } from '@mui/material';
+import { faUserPlus, faComment } from '@fortawesome/free-solid-svg-icons'
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+
+
+
+const FriendsList = (props: any) => {
+	const authCtx = useContext(AuthContext);
+	const [snackbarOpen, setSnackbarOpen] = useState(false);
+
+	const currentUserId = authCtx.userId;
+
+	const handleCloseSnackbar = () => {
+		setSnackbarOpen(false);
+	}
+
+	const handleDemand = async (event: FormEvent, receiverId: number) => {
+		event.preventDefault();
+		try {
+			const response = await fetch(`http://localhost:3000/friendship/create`,{
+				method: 'POST',
+				headers: {
+				  'Content-Type': 'application/json',
+				  Authorization: `Bearer ${authCtx.token}`,
+				},
+				body: JSON.stringify({requesterId: currentUserId, receiverId: receiverId}),
+			});
+			await response.json();
+			setSnackbarOpen(true);
+			if (!response.ok) {
+				console.log("POST error on /friendship/create");
+				return "error";
+			  }
+		} catch (error) {
+			console.log("error", error);
+		  }
+		}
+
+	return (
+		<>
+			<h2 className="title-users">{props.title} - <span>{props.users.length}</span></h2>
+			{/* <List> */}
+				{props.users.map(user => (
+					<ListItem key={user.id} className='users-list'>
+						<div className='avatar-username-inlist'>
+							<ListItemAvatar>
+								<MyAvatar style="s" authCtx={authCtx} alt={"avatar"} avatar={user.avatar} ftAvatar={user.ftAvatar} id={user.id} />
+							</ListItemAvatar>
+							<Link to={`/users/profile/${user.id}`} className="profile-link">{user.username}</Link>
+						</div>
+						<br />
+						<div className='btn-inlist'>
+							<FontAwesomeIcon icon={faUserPlus} onClick={(event: FormEvent) => { handleDemand(event, user.id) }} className='add-friend'/>
+							<Snackbar
+								anchorOrigin={{ vertical: 'top', horizontal: 'center' }}
+								open={snackbarOpen}
+								autoHideDuration={1000}
+								onClose={handleCloseSnackbar}
+								message="Request sent"
+								/>
+							<Link to="/chat/message"> <FontAwesomeIcon icon={faComment} className='add-friend'/></Link>
+						</div>
+					</ListItem>
+				))}
+			{/* </List> */}
+		</>
+	  );
+
+}
+
+export default FriendsList;
