+++ Front-End/src/index.css
@@ -15,7 +15,6 @@
 }
 
 a {
-	font-weight: 500;
 	color: #646cff;
 	text-decoration: inherit;
 }
