+++ Front-End/src/style/PopUpChannel.css
@@ -5,7 +5,7 @@
     width: 100%;
     height: 100%;
     background-color: rgba(0, 0, 0, 0.5);
-    z-index: 1;
+    z-index: 10;
 }
 
 .global-popup {
@@ -36,5 +36,5 @@
     justify-content: center;
     align-items: center;
 }
-  
+
