+++ Front-End/src/style/utils.css
@@ -5,18 +5,26 @@
 }
 
 .add-friend {
-    color: #699BF7;
-    font-size: 35px;
-	/* width: fit-content; */
+	color: #699bf7;
 	cursor: pointer;
-	margin-left: 13px;
-    margin-bottom: 7px;
-}
+	height: 20px;
+	/* color: black; */
+	padding: 11px;
+	border: solid 1px #80808024;
+	border-radius: 8px;
+	width: 25px;
+	margin-left: 10px;
+	margin-right: 10px;
+	}
 
 .add-friend:hover {
-    color: #C7B9FF;
+    /* color: #C7B9FF; */
+	background-color: #c7b9ff1f;
+    box-shadow: 0px 1px 1px #8b8a8a;
 }
 
 .profile-link {
-	color: black
+	color: black;
+	font-weight: 300;
 }
+
